Code refactoring: separate schema validation.
